requirements.txt
python-telegram-bot==21.6
httpx==0.27.2
qiskit==1.2.4
qiskit-aer==0.15.1
qiskit-ibm-runtime==0.23.0
amazon-braket-sdk==1.75.1
boto3==1.34.162
openai==1.46.0
anthropic==0.34.2
